// src/payload/payload-types.ts
export interface extras {
  name?: string | null
  price?: number | null
  id?: string | null
}

export interface menuItem {
  name: string | null
  image?: string | Media | null
  description: string | null
  base_price: number | null
  sizes: extras[]
  extraIngredientPrices: extras[]
}
export interface Media {
  id: string
  alt: string
  caption?:
  | {
    [k: string]: unknown
  }[]
  | null
  updatedAt: string
  createdAt: string
  url?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
}